+++ spyrit/core/recon.py
@@ -56,7 +56,11 @@ class PseudoInverse(nn.Module):
 
             :attr:`meas_op`: Measurement operator. Any class that
             implements a :meth:`pinv` method can be used, e.g.,
-            :class:`~spyrit.core.forwop.HadamSplit`.
+            :class:`~spyrit.core.meas.HadamSplit`.
+
+            :attr:`kwargs`: Additional keyword arguments that are passed to
+            the :meth:`pinv` method of the measurement operator. Can be used
+            to specify a regularization parameter.
 
         Shape:
 
@@ -85,20 +89,27 @@ class TikhonovMeasurementPriorDiag(nn.Module):
     r"""
     Tikhonov regularisation with prior in the measurement domain.
 
-    Considering linear measurements :math:`m = Hx \in\mathbb{R}^M`, where :math:`H = GF` is the measurement matrix and :math:`x\in\mathbb{R}^N` is a vectorized image, it estimates :math:`x` from :math:`m` by approximately minimizing
+    Considering linear measurements :math:`m = Hx \in\mathbb{R}^M`, where
+    :math:`H = GF` is the measurement matrix and :math:`x\in\mathbb{R}^N` is a
+    vectorized image, it estimates :math:`x` from :math:`m` by approximately
+    minimizing
 
     .. math::
         \|m - GFx \|^2_{\Sigma^{-1}_\alpha} + \|F(x - x_0)\|^2_{\Sigma^{-1}}
 
-    where :math:`x_0\in\mathbb{R}^N` is a mean image prior, :math:`\Sigma\in\mathbb{R}^{N\times N}` is a covariance prior, and :math:`\Sigma_\alpha\in\mathbb{R}^{M\times M}` is the measurement noise covariance. The matrix :math:`G\in\mathbb{R}^{M\times N}` is a subsampling matrix.
+    where :math:`x_0\in\mathbb{R}^N` is a mean image prior,
+    :math:`\Sigma\in\mathbb{R}^{N\times N}` is a covariance prior, and
+    :math:`\Sigma_\alpha\in\mathbb{R}^{M\times M}` is the measurement noise
+    covariance. The matrix :math:`G\in\mathbb{R}^{M\times N}` is a
+    subsampling matrix.
 
     .. note::
-        The class is instantiated from :math:`\Sigma`, which represents the covariance of :math:`Fx`.
+        The class is instantiated from :math:`\Sigma`, which represents the
+        covariance of :math:`Fx`.
 
     Args:
         - :attr:`sigma`:  covariance prior with shape :math:`N` x :math:`N`
         - :attr:`M`: number of measurements :math:`M`
-
 
     Attributes:
         :attr:`comp`: The learnable completion layer initialized as
@@ -159,7 +170,6 @@ class TikhonovMeasurementPriorDiag(nn.Module):
             \hat{x} &= x_0 + F^{-1}\begin{bmatrix}\Sigma_1 \\ \Sigma_{21} \end{bmatrix}
                       [\Sigma_1 + \Sigma_\alpha]^{-1} (m - GF x_0)
 
-
         See Lemma B.0.5 of the PhD dissertation of A. Lorente Mur (2021):
         https://theses.hal.science/tel-03670825v1/file/these.pdf
 
@@ -199,28 +209,61 @@ class TikhonovMeasurementPriorDiag(nn.Module):
 # =============================================================================
 class Denoise_layer(nn.Module):
     # =========================================================================
-    r"""Wiener filter that assumes additive white Gaussian noise.
+    r"""Defines a learnable Wiener filter that assumes additive white Gaussian noise.
 
-    :math:`y = \sigma_\text{prior}^2/(\sigma^2_\text{prior} + \sigma^2_\text{meas}) x`,
-    where :math:`\sigma^2_\text{prior}` is the variance prior and
-    :math:`\sigma^2_\text{meas}` is the variance of the measurement, :math:`x`
-    is the input vector and :math:`y` is the output vector.
+    The filter is pre-defined upon initialization with the standard deviation prior
+    (if known), or with an integer representing the size of the input vector.
+    In the second case, the standard deviation prior is initialized at random
+    from a uniform (0,2/size) distribution.
+
+    Using the foward method (the implicit call method), the filter is fully
+    defined:
+
+    .. math::
+        \sigma_\text{prior}^2/(\sigma^2_\text{prior} + \sigma^2_\text{meas})
+
+    where :math:`\sigma^2_\text{prior}` is the variance prior defined at
+    initialization and :math:`\sigma^2_\text{meas}` is the measurement variance
+    defined using the forward method. The value given by the equation above
+    can then be multiplied by the measurement vector to obtain the denoised
+    measurement vector.
+
+    ..note::
+        The weight (defined at initialization or accessible through the
+        attribute :attr:`weight`) should not be squared (as it is squared when
+        the forward method is called).
 
     Args:
-        :attr:`M` (int): size of incoming vector
+        :attr:`std_dev_or_size` (torch.tensor or int): 1D tensor representing
+        the standard deviation prior or an integer defining the size of the
+        randomly-initialized standard deviation prior. If an array is passed
+        and it is not 1D, it is flattened. It is stored internally as a
+        :class:`nn.Parameter`, whose :attr:`data` attribute is accessed through
+        the :attr:`sigma` attribute, and whose :attr:`requires_grad` attribute
+        is accessed through the :attr:`requires_grad` attribute.
 
-    Shape:
-        - Input: :math:`(*, M)`.
-        - Output: :math:`(*, M)`.
+    Shape for forward call:
+        - Input: :math:`(*, in\_features)` measurement variance.
+        - Output: :math:`(*, in\_features)` fully defined Wiener filter.
 
     Attributes:
         :attr:`weight`:
         The learnable standard deviation prior :math:`\sigma_\text{prior}` of
-        shape :math:`(M, 1)`. The values are initialized from
-        :math:`\mathcal{U}(-\sqrt{k}, \sqrt{k})`, where :math:`k = 1/M`.
+        shape :math:`(in\_features, 1)`. The values are initialized from
+        :math:`\mathcal{U}(-\sqrt{k}, \sqrt{k})`, where :math:`k = 1/in\_features`.
+
+        :attr:`sigma`:
+        The learnable standard deviation prior :math:`\sigma_\text{prior}` of shape
+        :math:`(, in\_features)`. If the input is an integer, the standard deviation prior
+        is initialized at random from  :math:`\mathcal{U}(-\sqrt{k}, \sqrt{k})`,
+        where :math:`k = 1/in\_features`.
 
         :attr:`in_features`:
-        The number of input features equal to :math:`M`.
+        The number of input features.
+
+        :attr:`requires_grad`:
+        A boolean indicating whether the autograd should record operations on
+        the standard deviation tensor. Default is True.
 
     Example:
         >>> m = Denoise_layer(30)
@@ -230,40 +273,72 @@ class Denoise_layer(nn.Module):
         torch.Size([128, 30])
     """
 
-    def __init__(self, M: int):
+    def __init__(
+        self, std_dev_prior_or_size: Union[torch.tensor, int], requires_grad=True
+    ):
         super(Denoise_layer, self).__init__()
-        self.in_features = M
-        self.weight = nn.Parameter(torch.Tensor(M))
-        self.reset_parameters()
+
+        if isinstance(std_dev_prior_or_size, int):
+            self.weight = nn.Parameter(
+                torch.Tensor(std_dev_prior_or_size), requires_grad=requires_grad
+            )
+            self.reset_parameters()
+
+        else:
+            if not isinstance(std_dev_prior_or_size, torch.Tensor):
+                raise TypeError(
+                    "std_dev_or_size should be an integer or a torch.Tensor"
+                )
+            self.weight = nn.Parameter(
+                std_dev_prior_or_size.reshape(-1), requires_grad=requires_grad
+            )
+
+    @property
+    def in_features(self):
+        return self.weight.data.numel()
 
     def reset_parameters(self):
         r"""
         Resets the standard deviation prior :math:`\sigma_\text{prior}`.
 
         The values are initialized from :math:`\mathcal{U}(-\sqrt{k}, \sqrt{k})`,
-        where :math:`k = 1/M`. They are stored in the :attr:`weight` attribute.
+        where :math:`k = 1/in\_features`. They are stored in the :attr:`weight`
+        attribute.
         """
         nn.init.uniform_(self.weight, 0, 2 / math.sqrt(self.in_features))
 
-    def forward(self, inputs: torch.tensor) -> torch.tensor:
+    def forward(self, sigma_meas_squared: torch.tensor) -> torch.tensor:
         r"""
-        Applies a transformation to the incoming data: :math:`y = \sigma_\text{prior}^2/(\sigma_\text{prior}^2+x)`.
+        Fully defines the Wiener filter with the measurement variance.
 
-        :math:`x` is the input tensor (see :attr:`inputs`) and
-        :math:`\sigma_\text{prior}` is the standard deviation prior (see :attr:`self.weight`).
+        This outputs :math:`\sigma_\text{prior}^2/(\sigma_\text{prior}^2 + \sigma^2_\text{meas})`,
+        where :math:`\sigma^2_\text{meas}` is the measurement variance (see :attr:`sigma_meas_squared`) and
+        :math:`\sigma_\text{prior}` is the standard deviation prior defined
+        upon construction of the class (see :attr:`self.weight`).
+
+        ..note::
+            The measurement variance should be squared before being passed to
+            this method, unlike the standard deviation prior (defined at construction).
 
         Args:
-            :attr:`inputs` (torch.tensor): input tensor :math:`x` of shape
-            :math:`(N, *, in\_features)`
+            :attr:`sigma_meas_squared` (torch.tensor): input tensor :math:`\sigma^2_\text{meas}`
+            of shape :math:`(*, in\_features)`
 
         Returns:
-            torch.tensor: The transformed data :math:`y` of shape
-            :math:`(N, in\_features)`
+            torch.tensor: The multiplicative filter of shape
+            :math:`(*, in\_features)`
 
         Shape:
-
+            - Input: :math:`(*, in\_features)`
+            - Output: :math:`(*, in\_features)`
         """
-        return self.tikho(inputs, self.weight)
+        if sigma_meas_squared.shape[-1] != self.in_features:
+            raise ValueError(
+                "The last dimension of the input tensor "
+                + f"({sigma_meas_squared.shape[-1]})should be equal to the number of "
+                + f"input features ({self.in_features})."
+            )
+        return self.tikho(sigma_meas_squared, self.weight)
 
     def extra_repr(self):
         return "in_features={}".format(self.in_features)
@@ -819,227 +894,226 @@ class PositiveParameters(nn.Module):
 
 
 # =============================================================================
-class LearnedPGD(nn.Module):
-    r"""Learned Proximal Gradient Descent reconstruction network.
-    Iterative algorithm that alternates between a gradient step and a proximal step,
-    where the proximal operator is learned denoiser. The update rule is given by:
+# class LearnedPGD(nn.Module):
+#     r"""Learned Proximal Gradient Descent reconstruction network.
+#     Iterative algorithm that alternates between a gradient step and a proximal step,
+#     where the proximal operator is learned denoiser. The update rule is given by:
 
-    :math:`x_{k+1} = prox(\hat{x_k} - step * H^T (Hx_k - y))=
-    denoi(\hat{x_k} - step * H^T (Hx_k - y))`
+#     :math:`x_{k+1} = prox(\hat{x_k} - step * H^T (Hx_k - y))=
+#     denoi(\hat{x_k} - step * H^T (Hx_k - y))`
 
-    Args:
-        :attr:`noise`: Acquisition operator (see :class:`~spyrit.core.noise`)
+#     Args:
+#         :attr:`noise`: Acquisition operator (see :class:`~spyrit.core.noise`)
 
-        :attr:`prep`: Preprocessing operator (see :class:`~spyrit.core.prep`)
+#         :attr:`prep`: Preprocessing operator (see :class:`~spyrit.core.prep`)
 
-        :attr:`denoi` (optional): Image denoising operator
-        (see :class:`~spyrit.core.nnet`).
-        Default :class:`~spyrit.core.nnet.Identity`
+#         :attr:`denoi` (optional): Image denoising operator
+#         (see :class:`~spyrit.core.nnet`).
+#         Default :class:`~spyrit.core.nnet.Identity`
 
-        :attr:`iter_stop` (int): Number of iterations of the LPGD algorithm
-        (commonly 3 to 10, trade-off between accuracy and speed).
-        Default 3 (for speed and with higher accuracy than post-processing denoising)
+#         :attr:`iter_stop` (int): Number of iterations of the LPGD algorithm
+#         (commonly 3 to 10, trade-off between accuracy and speed).
+#         Default 3 (for speed and with higher accuracy than post-processing denoising)
 
-        :attr:`step` (float): Step size of the LPGD algorithm. Default is None,
-        and it is estimated as the inverse of the Lipschitz constant of the gradient of the
-        data fidelity term.
-            - If :math:`meas_op.N` is available, the step size is estimated as
-            :math:`step=1/L=1/\text{meas_op.N}`, true for Hadamard operators.
-            - If not, the step size is estimated from by computing
-            the Lipschitz constant as the largest singular value of the
-            Hessians, :math:`L=\lambda_{\max}(H^TH)`. If this fails,
-            the step size is set to 1e-4.
+#         :attr:`step` (float): Step size of the LPGD algorithm. Default is None,
+#         and it is estimated as the inverse of the Lipschitz constant of the gradient of the
+#         data fidelity term.
+#             - If :math:`meas_op.N` is available, the step size is estimated as
+#             :math:`step=1/L=1/\text{meas_op.N}`, true for Hadamard operators.
+#             - If not, the step size is estimated from by computing
+#             the Lipschitz constant as the largest singular value of the
+#             Hessians, :math:`L=\lambda_{\max}(H^TH)`. If this fails,
+#             the step size is set to 1e-4.
 
-        :attr:`step_estimation` (bool): Default False. See :attr:`step` for details.
+#         :attr:`step_estimation` (bool): Default False. See :attr:`step` for details.
 
-        :attr:`step_grad` (bool): Default False. If True, the step size is learned
-        as a parameter of the network. Not tested yet.
+#         :attr:`step_grad` (bool): Default False. If True, the step size is learned
+#         as a parameter of the network. Not tested yet.
 
-        :attr:`wls` (bool): Default False. If True, the data fidelity term is
-        modified to be the weighted least squares (WLS) term, which approximates
-        the Poisson likelihood. In this case, the data fidelity term is
-        :math:`\|Hx-y\|^2_{C^{-1}}`, where :math:`C` is the covariance matrix.
-        We assume that :math:`C` is diagonal, and the diagonal elements are
-        the measurement noise variances, estimated from :class:`~spyrit.core.prep.sigma`.
+#         :attr:`wls` (bool): Default False. If True, the data fidelity term is
+#         modified to be the weighted least squares (WLS) term, which approximates
+#         the Poisson likelihood. In this case, the data fidelity term is
+#         :math:`\|Hx-y\|^2_{C^{-1}}`, where :math:`C` is the covariance matrix.
+#         We assume that :math:`C` is diagonal, and the diagonal elements are
+#         the measurement noise variances, estimated from :class:`~spyrit.core.prep.sigma`.
 
-        :attr:`gt` (torch.tensor): Ground-truth images. If available, the mean
-        squared error (MSE) is computed and logged. Default None.
+#         :attr:`gt` (torch.tensor): Ground-truth images. If available, the mean
+#         squared error (MSE) is computed and logged. Default None.
 
-        :attr:`log_fidelity` (bool): Default False. If True, the data fidelity term
-        is logged for each iteration of the LPGD algorithm.
+#         :attr:`log_fidelity` (bool): Default False. If True, the data fidelity term
+#         is logged for each iteration of the LPGD algorithm.
 
-    Input / Output:
-        :attr:`input`: Ground-truth images with shape :math:`(B,C,H,W)`
+#     Input / Output:
+#         :attr:`input`: Ground-truth images with shape :math:`(B,C,H,W)`
 
-        :attr:`output`: Reconstructed images with shape :math:`(B,C,H,W)`
+#         :attr:`output`: Reconstructed images with shape :math:`(B,C,H,W)`
 
-    Attributes:
-        :attr:`Acq`: Acquisition operator initialized as :attr:`noise`
+#     Attributes:
+#         :attr:`Acq`: Acquisition operator initialized as :attr:`noise`
 
-        :attr:`prep`: Preprocessing operator initialized as :attr:`prep`
+#         :attr:`prep`: Preprocessing operator initialized as :attr:`prep`
 
-        :attr:`pinv`: Analytical reconstruction operator initialized as
-        :class:`~spyrit.core.recon.PseudoInverse()`
+#         :attr:`pinv`: Analytical reconstruction operator initialized as
+#         :class:`~spyrit.core.recon.PseudoInverse()`
 
-        :attr:`Denoi`: Image denoising operator initialized as :attr:`denoi`
+#         :attr:`Denoi`: Image denoising operator initialized as :attr:`denoi`
 
-    Example:
-        >>> B, C, H, M = 10, 1, 64, 64**2
-        >>> Ord = torch.ones((H,H))
-        >>> meas = HadamSplit(M, H, Ord)
-        >>> noise = NoNoise(meas)
-        >>> prep = SplitPoisson(1.0, M, H*H)
-        >>> denoi = denoi = Unet()
-        >>> recnet = LearnedPGD(noise, prep, denoi)
-        >>> x = torch.FloatTensor(B,C,H,H).uniform_(-1, 1)
-        >>> z = recnet(x)
-        >>> print(z.shape)
-        >>> print(torch.linalg.norm(x - z)/torch.linalg.norm(x))
-        torch.Size([10, 1, 64, 64])
-        tensor(5.8912e-06)
-    """
+#     Example:
+#         >>> B, C, H, M = 10, 1, 64, 64**2
+#         >>> Ord = torch.ones((H,H))
+#         >>> meas = HadamSplit(M, H, Ord)
+#         >>> noise = NoNoise(meas)
+#         >>> prep = SplitPoisson(1.0, M, H*H)
+#         >>> recnet = LearnedPGD(noise, prep)
+#         >>> x = torch.FloatTensor(B,C,H,H).uniform_(-1, 1)
+#         >>> z = recnet(x)
+#         >>> print(z.shape)
+#         torch.Size([10, 1, 64, 64])
+#         >>> print(torch.linalg.norm(x - z)/torch.linalg.norm(x))
+#         tensor(5.8912e-06)
+#     """
 
-    def __init__(
-        self,
-        noise,
-        prep,
-        denoi=nn.Identity(),
-        iter_stop=3,
-        x0=0,
-        step=None,
-        step_estimation=False,
-        step_grad=False,
-        step_decay=1,
-        wls=False,
-        gt=None,
-        log_fidelity=False,
-        res_learn=False,
-    ):
-        super().__init__()
-        # nn.module
-        self.acqu = noise
-        self.prep = prep
-        self.denoi = denoi
+#     def __init__(
+#         self,
+#         noise,
+#         prep,
+#         denoi=nn.Identity(),
+#         iter_stop=3,
+#         x0=0,
+#         step=None,
+#         step_estimation=False,
+#         step_grad=False,
+#         step_decay=1,
+#         wls=False,
+#         gt=None,
+#         log_fidelity=False,
+#         res_learn=False,
+#     ):
+#         super().__init__()
+#         # nn.module
+#         self.acqu = noise
+#         self.prep = prep
+#         self.denoi = denoi
 
-        self.pinv = PseudoInverse()
+#         self.pinv = PseudoInverse()
 
-        # LPGD algo
-        self.x0 = x0
-        self.iter_stop = iter_stop
-        self.step = step
-        self.step_estimation = step_estimation
-        self.step_grad = step_grad
-        self.step_decay = step_decay
-        self.res_learn = res_learn
+#         # LPGD algo
+#         self.x0 = x0
+#         self.iter_stop = iter_stop
+#         self.step = step
+#         self.step_estimation = step_estimation
+#         self.step_grad = step_grad
+#         self.step_decay = step_decay
+#         self.res_learn = res_learn
 
-        # Init step size (estimate)
-        self.set_stepsize(step)
+#         # Init step size (estimate)
+#         self.set_stepsize(step)
 
-        # WLS
-        self.wls = wls
+#         # WLS
+#         self.wls = wls
 
-        # Log fidelity
-        self.log_fidelity = log_fidelity
+#         # Log fidelity
+#         self.log_fidelity = log_fidelity
 
-        # Log MSE (Ground truth available)
-        if gt is not None:
-            self.x_gt = nn.Parameter(
-                torch.tensor(gt.reshape(gt.shape[0], -1)), requires_grad=False
-            )
-        else:
-            self.x_gt = None
+#         # Log MSE (Ground truth available)
+#         if gt is not None:
+#             self.x_gt = nn.Parameter(
+#                 torch.tensor(gt.reshape(gt.shape[0], -1)), requires_grad=False
+#             )
+#         else:
+#             self.x_gt = None
 
-    def step_schedule(self, step):
-        if self.step_decay != 1:
-            step = [step * self.step_decay**i for i in range(self.iter_stop)]
-        elif self.iter_stop > 1:
-            step = [step for i in range(self.iter_stop)]
-        else:
-            step = [step]
-        return step
+#     def step_schedule(self, step):
+#         if self.step_decay != 1:
+#             step = [step * self.step_decay**i for i in range(self.iter_stop)]
+#         elif self.iter_stop > 1:
+#             step = [step for i in range(self.iter_stop)]
+#         else:
+#             step = [step]
+#         return step
 
-    def set_stepsize(self, step):
-        if step is None:
-            # Stimate stepsize from Lipschitz constant
-            if hasattr(self.acqu.meas_op, "N"):
-                step = 1 / self.acqu.meas_op.N
-            else:
-                # Estimate step size as 1/sv_max(H^TH); if failed, set to 1e-4
-                self.step_estimation = True
-                step = 1e-4
+#     def set_stepsize(self, step):
+#         if step is None:
+#             # Stimate stepsize from Lipschitz constant
+#             if hasattr(self.acqu.meas_op, "N"):
+#                 step = 1 / self.acqu.meas_op.N
+#             else:
+#                 # Estimate step size as 1/sv_max(H^TH); if failed, set to 1e-4
+#                 self.step_estimation = True
+#                 step = 1e-4
 
-        step = self.step_schedule(step)
-        # step = nn.Parameter(torch.tensor(step), requires_grad=self.step_grad)
-        step = PositiveParameters(step, requires_grad=self.step_grad)
-        self.step = step
+#         step = self.step_schedule(step)
+#         # step = nn.Parameter(torch.tensor(step), requires_grad=self.step_grad)
+#         step = PositiveParameters(step, requires_grad=self.step_grad)
+#         self.step = step
 
-    def forward(self, x):
-        r"""Full pipeline of reconstrcution network
+#     def forward(self, x):
+#         r"""Full pipeline of reconstrcution network
 
-        Args:
-            :attr:`x`: ground-truth images
+#         Args:
+#             :attr:`x`: ground-truth images
 
-        Shape:
-            :attr:`x`: ground-truth images with shape :math:`(B,C,H,W)`
+#         Shape:
+#             :attr:`x`: ground-truth images with shape :math:`(B,C,H,W)`
 
-            :attr:`output`: reconstructed images with shape :math:`(B,C,H,W)`
+#             :attr:`output`: reconstructed images with shape :math:`(B,C,H,W)`
 
-        Example:
-            >>> B, C, H, M = 10, 1, 64, 64**2
-            >>> Ord = torch.ones((H,H))
-            >>> meas = HadamSplit(M, H, Ord)
-            >>> noise = NoNoise(meas)
-            >>> prep = SplitPoisson(1.0, M, H*H)
-            >>> recnet = LearnedPGD(noise, prep)
-            >>> x = torch.FloatTensor(B,C,H,H).uniform_(-1, 1)
-            >>> z = recnet(x)
-            >>> print(z.shape)
-            >>> print(torch.linalg.norm(x - z)/torch.linalg.norm(x))
-            torch.Size([10, 1, 64, 64])
-            tensor(5.8912e-06)
-        """
+#         Example:
+#             >>> B, C, H, M = 10, 1, 64, 64**2
+#             >>> Ord = torch.ones((H,H))
+#             >>> meas = HadamSplit(M, H, Ord)
+#             >>> noise = NoNoise(meas)
+#             >>> prep = SplitPoisson(1.0, M, H*H)
+#             >>> recnet = LearnedPGD(noise, prep)
+#             >>> x = torch.FloatTensor(B,C,H,H).uniform_(-1, 1)
+#             >>> z = recnet(x)
+#             >>> print(z.shape)
+#             torch.Size([10, 1, 64, 64])
+#             >>> print(torch.linalg.norm(x - z)/torch.linalg.norm(x))
+#             tensor(5.8912e-06)
+#         """
 
-        b, c, _, _ = x.shape
+#         b, c, _, _ = x.shape
 
-        # Acquisition
-        x = x.reshape(b * c, self.acqu.meas_op.N)  # shape x = [b*c,h*w] = [b*c,N]
-        x = self.acqu(x)  # shape x = [b*c, 2*M]
+#         # Acquisition
+#         x = x.reshape(b * c, self.acqu.meas_op.N)  # shape x = [b*c,h*w] = [b*c,N]
+#         x = self.acqu(x)  # shape x = [b*c, 2*M]
 
-        # Reconstruction
-        x = self.reconstruct(x)  # shape x = [bc, 1, h,w]
-        x = x.reshape(b, c, self.acqu.meas_op.h, self.acqu.meas_op.w)
+#         # Reconstruction
+#         x = self.reconstruct(x)  # shape x = [bc, 1, h,w]
+#         x = x.reshape(b, c, self.acqu.meas_op.h, self.acqu.meas_op.w)
 
-        return x
+#         return x
 
-    def acquire(self, x):
-        r"""Simulate data acquisition
+#     def acquire(self, x):
+#         r"""Simulate data acquisition
 
-        Args:
-            :attr:`x`: ground-truth images
+#         Args:
+#             :attr:`x`: ground-truth images
 
-        Shape:
-            :attr:`x`: ground-truth images with shape :math:`(B,C,H,W)`
+#         Shape:
+#             :attr:`x`: ground-truth images with shape :math:`(B,C,H,W)`
 
-            :attr:`output`: reconstructed images with concatenated noise level map with shape :math:`(BC,2,H,W)`
-        """
+#             :attr:`output`: reconstructed images with concatenated noise level map with shape :math:`(BC,2,H,W)`
+#         """
 
-        b, c, h, w = x.shape
-        x = 0.5 * (x + 1)
-        x = torch.cat((x, self.noise_level.expand(b, 1, h, w)), dim=1)
-        return x
+#         b, c, h, w = x.shape
+#         x = 0.5 * (x + 1)
+#         x = torch.cat((x, self.noise_level.expand(b, 1, h, w)), dim=1)
+#         return x
 
-    def set_noise_level(self, noise_level):
-        r"""Reset noise level value
+#     def set_noise_level(self, noise_level):
+#         r"""Reset noise level value
 
-        Args:
-            :attr:`noise_level`: noise level value in the range [0, 255]
+#         Args:
+#             :attr:`noise_level`: noise level value in the range [0, 255]
 
-        Shape:
-            :attr:`noise_level`: float value noise level :math:`(1)`
+#         Shape:
+#             :attr:`noise_level`: float value noise level :math:`(1)`
 
-            :attr:`output`: noise level tensor with shape :math:`(1)`
-        """
-        self.noise_level = torch.FloatTensor([noise_level / 255.0])
+#             :attr:`output`: noise level tensor with shape :math:`(1)`
+#         """
+#         self.noise_level = torch.FloatTensor([noise_level / 255.0])
 
 
 # %%===========================================================================================
@@ -1224,12 +1298,12 @@ class LearnedPGD(nn.Module):
             >>> meas = HadamSplit(M, H, Ord)
             >>> noise = NoNoise(meas)
             >>> prep = SplitPoisson(1.0, M, H*H)
-            >>> recnet = PinvNet(noise, prep)
+            >>> recnet = LearnedPGD(noise, prep)
             >>> x = torch.FloatTensor(B,C,H,H).uniform_(-1, 1)
             >>> z = recnet(x)
             >>> print(z.shape)
-            >>> print(torch.linalg.norm(x - z)/torch.linalg.norm(x))
             torch.Size([10, 1, 64, 64])
+            >>> print(torch.linalg.norm(x - z)/torch.linalg.norm(x))
             tensor(5.8912e-06)
         """